requirements.txt
ollama
pandas
chromadb
fastapi
python-multipart
uvicorn[standard]
python-dotenv
streamlit
pytest
openpyxl
matplotlib
